Sets a callback to be notified when the socket is actually closed.
When there are no more pending operations on the socket, if a close was requested, a callback will be called. In that callback we emit the "close" event.

Because the close action is "queued", after receiving the "close" event, we could receive "data" events, for instance, for read completions. Now, after a "close" we won't receive any events.