High resolution timer.
The function process.hrtime() is added, which provides a binding for
libuv's uv_hrtime. It is a number in nanoseconds that is not subjected to
changes in the system clock.

Since the high resolution timer is a 64 bit value and in Lua we have only
53 bits of integer precision, we represent it as a userdata large enough
to hold the value in it.

Some operations are provideed with these values: substraction (but not
addition), some relational operations (<, <=, >, >=) and extractor
methods. The value can be converted to microseconds, milliseconds or split
into seconds and nanoseconds.